Volume-1-Foundations/Code/requirements.txt
# AI for Networking Engineers - Volume 1: Foundations
# Python Dependencies

# Core AI APIs
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP and Networking
requests>=2.31.0
httpx>=0.25.0

# Network Automation
netmiko>=4.3.0
napalm>=4.1.0
textfsm>=1.1.3
ntc-templates>=4.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Parsing and Validation
jsonschema>=4.20.0
pyyaml>=6.0.1
jinja2>=3.1.2

# Caching and Storage
redis>=5.0.0
diskcache>=5.6.3

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Logging and Monitoring
structlog>=23.2.0
colorama>=0.4.6

# Type Checking and Linting
mypy>=1.7.0
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0

# Cloud/Secrets Management (Optional)
boto3>=1.34.0  # AWS Secrets Manager
azure-keyvault>=4.2.0  # Azure Key Vault
google-cloud-secret-manager>=2.17.0  # Google Secret Manager

# Utilities
click>=8.1.7  # CLI framework
rich>=13.7.0  # Terminal formatting
tabulate>=0.9.0  # Table formatting
tqdm>=4.66.0  # Progress bars
